requirements.txt
Flask
Flask-Cors
dnspython
urllib3
werkzeug
gunicorn
celery
redis